Moved tzrange into _common, gave it its own fold-aware fromutc().
